# reports/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageType": {
        "enum": [
          "text",
          "price_negotiation",
          "task_specification",
          "counter_offer",
          "accept",
          "reject"
        ],
        "title": "MessageType",
        "type": "string"
      },
      "NegotiatePriceRequest": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "expires_hours": {
            "default": 24,
            "title": "Expires Hours",
            "type": "integer"
          },
          "initiator": {
            "title": "Initiator",
            "type": "string"
          },
          "responder": {
            "title": "Responder",
            "type": "string"
          },
          "service_id": {
            "default": "",
            "title": "Service Id",
            "type": "string"
          }
        },
        "required": [
          "initiator",
          "responder",
          "amount"
        ],
        "title": "NegotiatePriceRequest",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "default": "",
            "title": "Body",
            "type": "string"
          },
          "message_type": {
            "$ref": "#/components/schemas/MessageType"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "sender": {
            "title": "Sender",
            "type": "string"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "required": [
          "sender",
          "recipient",
          "message_type"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key passed directly in the X-API-Key header (alternative to Bearer).",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "API key passed as Bearer token in the Authorization header.",
        "scheme": "bearer",
        "type": "http"
      },
      "X402Payment": {
        "description": "x402 payment proof for stateless authentication.",
        "in": "header",
        "name": "X-PAYMENT",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "A2A Commerce Gateway",
    "version": "1.2.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/messaging/messages": {
      "get": {
        "operationId": "get_messages_v1_messaging_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thread_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Thread Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Messages",
        "tags": [
          "messaging"
        ]
      },
      "post": {
        "operationId": "send_message_v1_messaging_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "messaging"
        ]
      }
    },
    "/v1/messaging/negotiations": {
      "post": {
        "operationId": "negotiate_price_v1_messaging_negotiations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiatePriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Negotiate Price",
        "tags": [
          "messaging"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    },
    {
      "X402Payment": []
    }
  ]
}